requirements.txt
pypdfium2~=4.20.0
Pillow~=10.0.1
transformers~=4.33.1
pytesseract
tensorflow